mcpx/dsp: Move some debug functions out to debug.c
